+++ examples/west_coast_random.py
@@ -1,16 +1,19 @@
 """
 .. module:: west_coast_random
     :platform: Windows
-    :synopsis: Example code making a scenario in west_coast_usa and having two
-               cars drive around randomly.
+    :synopsis: Example code making a scenario in west_coast_usa and having a
+               car drive around randomly.
 
 .. moduleauthor:: Marc Müller <mmueller@beamng.gmbh>
 
 """
 import mmap
 import random
+import sys
 
 from time import sleep
+
+import numpy as np
 
 from matplotlib import pyplot as plt
 from matplotlib.pyplot import imshow
@@ -37,58 +40,46 @@ def main():
 
     plt.ion()
 
-    beamng = BeamNGpy('localhost', 64256)
+    beamng = BeamNGpy('localhost', 64256,
+                      home='D:\\beamng.research_unlimited.v1.3.0.0')
 
     # Create a scenario in west_coast_usa
     scenario = Scenario('west_coast_usa', 'research_test',
                         description='Random driving for research')
 
-    # Set up first vehicle, with a frontal camera, gforces sensor, and lidar.
+    # Set up first vehicle, with two cameras, gforces sensor, lidar, electrical
+    # sensors, and damage sensors
     vehicle = Vehicle('ego_vehicle', model='etk800',
-                      licence='FOO', color='Red')
+                      licence='RED', color='Red')
 
     # Set up sensors
     pos = (-0.3, 1, 1.0)
     direction = (0, 1, 0)
     fov = 120
     resolution = (512, 512)
-    camera = Camera(pos, direction, fov, resolution,
-                    colour=True, depth=True, annotation=True)
-    vehicle.attach_sensor('front_cam', camera)
-    gforces = GForces()
-    vehicle.attach_sensor('gforces', gforces)
-    lidar = Lidar()
-    vehicle.attach_sensor('lidar', lidar)
-
-    # Add the vehicle to the scenario
-    scenario.add_vehicle(vehicle, pos=(-717.121, 101, 118.675),
-                         rot=(0, 0, 45))
-
-    # Set up the second vehicle, with a frontal camera, gforces sensor,
-    # electrics, and damage sensors.
-    other_vehicle = Vehicle('other_vehicle', model='etk800',
-                            licence='BAR', color='Blue')
-    # Set up sensors
-    pos = (-0.3, 1, 1.0)
-    direction = (0, 1, 0)
-    fov = 60
+    front_camera = Camera(pos, direction, fov, resolution,
+                          colour=True, depth=True, annotation=True)
+    pos = (0.0, 3, 1.0)
+    direction = (0, -1, 0)
+    fov = 90
     resolution = (512, 512)
-    camera = Camera(pos, direction, fov, resolution,
-                    colour=True, depth=True, annotation=True)
+    back_camera = Camera(pos, direction, fov, resolution,
+                         colour=True, depth=True, annotation=True)
 
     gforces = GForces()
     electrics = Electrics()
     damage = Damage()
+    lidar = Lidar()
 
     # Attach them
-    other_vehicle.attach_sensor('front_cam', camera)
-    other_vehicle.attach_sensor('gforces', gforces)
-    other_vehicle.attach_sensor('electrics', electrics)
-    other_vehicle.attach_sensor('damage', damage)
+    vehicle.attach_sensor('front_cam', front_camera)
+    vehicle.attach_sensor('back_cam', back_camera)
+    vehicle.attach_sensor('gforces', gforces)
+    vehicle.attach_sensor('lidar', lidar)
+    vehicle.attach_sensor('electrics', electrics)
+    vehicle.attach_sensor('damage', damage)
 
-    # Add vehicle to scenario
-    scenario.add_vehicle(other_vehicle, pos=(-722, 101, 118.675),
-                         rot=(0, 0, 45))
+    scenario.add_vehicle(vehicle, pos=(-717.121, 101, 118.675), rot=(0, 0, 45))
 
     # Compile the scenario and place it in BeamNG's map folder
     scenario.make(beamng)
@@ -107,35 +98,28 @@ def main():
         bng.hide_hud()
 
         assert vehicle.skt
-        assert other_vehicle.skt
 
-        # Send random inputs to both vehices and advance the simulation 6 steps
+        # Send random inputs to vehice and advance the simulation 20 steps
         for _ in range(1024):
             throttle = random.uniform(0.0, 1.0)
             steering = random.uniform(-1.0, 1.0)
             brake = random.choice([0, 0, 0, 1])
             vehicle.control(throttle=throttle, steering=steering, brake=brake)
 
-            throttle = random.uniform(0.0, 1.0)
-            steering = random.uniform(-1.0, 1.0)
-            brake = random.choice([0, 0, 0, 1])
-            other_vehicle.control(throttle=throttle, steering=steering,
-                                  brake=brake)
-
-            bng.step(6)
+            bng.step(20)
 
             # Retrieve sensor data and show the camera data.
             sensors = bng.poll_sensors(vehicle)
+
             a_colour.imshow(sensors['front_cam']['colour'].convert('RGB'))
             a_depth.imshow(sensors['front_cam']['depth'].convert('L'))
             a_annot.imshow(sensors['front_cam']['annotation'].convert('RGB'))
 
-            sensors = bng.poll_sensors(other_vehicle)
-            b_colour.imshow(sensors['front_cam']['colour'].convert('RGB'))
-            b_depth.imshow(sensors['front_cam']['depth'].convert('L'))
-            b_annot.imshow(sensors['front_cam']['annotation'].convert('RGB'))
+            b_colour.imshow(sensors['back_cam']['colour'].convert('RGB'))
+            b_depth.imshow(sensors['back_cam']['depth'].convert('L'))
+            b_annot.imshow(sensors['back_cam']['annotation'].convert('RGB'))
 
-            plt.pause(0.0016)
+            plt.pause(0.00001)
     finally:
         bng.close()